Update call to AssembleForLaunch for KSP 0.21
KSP 0.21 changed the parameters of ShipConstruction.AssembleForLaunch: the
state variable is now of type "Game" and it now takes a crew manifest list.

I have used HighLogic.CurrentGame for the game state (I do NOT know if this
is correct, but during my testing, I didn't notice any problems I would
attribute to this).

I used "null" for the crew manifest. This seems to work nicely and as a
side-effect, it fixes the launch-pad's magical ability to create Kerbals:
crewed ships are now built without a crew.