Remove the hard-coded RocketParts density.
Having 1t/u for the density of RocketParts is very harsh when 1m^3 can hold
around 200u of anything else. This change doesn't actually change the
density of rocket parts, but it does make it so changing the density will
cause the right amount of rocket parts to be used when building a ship.