Clear the vessel's Landed flag.
It seems the ship being forced to ground is due to the landed flag being
true when physics kicks in. This breaks the landedAt setting, but oh well.

The explosions are fixed (the craft now sits nicely on the pad when
SpawnHeightOffset is set appropriately), but there are still problems with
control input.