Up the smelter's rate a little and add a small smelter.
The rate of the smelter is now 3.484u/s instead of 3 (there's actually a
reason for the funny number). This puts the rate at just slower than the
auger (change not yet committed).

Also, add a half-size (by mass. ~80% by scale): half the mass and half the
rate, 75% of the power (smaller mass to surface area ratio: more heat
loss).