Use MODEL nodes for the recycling bin.
Also, correct the swapping of the large and small bins.

It turns out that "mesh" is deprecated because it doesn't actually do what
it says. This fixes the small recycling bin getting the large bin's model.

NOTE: in this version, the EL directory in GameData *MUST* be
"Extraplanetary Launchpads" or the bins will not load.