Prevent the event catcher module from multiplying.
It doesn't seem to be possible to hook into the craft save process in order
to prevent a module from being saved, so butcher the config node such that
KSP will ignore it. Modules not defined in the original part config node
cannot be loaded, so "multiplying" isn't really correct, but should that
limitation ever be removed, it will become correct.