Rework craft spawning to avoid twisted ships.
Many thanks to angavrilov for his tips. It turns out I need to wait for the
parts to initialize properly before docking the vessel to the pad. There
are still some issues (not warping through the spawn causes not-so-rapid
unplanned disassembly, warping though the spawn avoids that but the orbit
renderer seems to be messed up), but this is good progress towards fixing
issue #3.