Extend the airlock collider to hit the main collider.
Many thanks to 5thHorseman for the tip that the only (stock?) part with an
easy to use hatch is the science lab. It turns out that the lab's airlock
collider is well and truly embedded in the main collider and it turns out
that KSP seems to somehow use the intersection to find out if the mouse is
"over" the airlock. KSP does something weird such that the airlock isn't
quite in the expected location, but it is now nice and easy to find the
airlock with the mouse.

Fixes #29.