Make workshops non-functional on probe cores.
This is really a workaround for Module Manager not having a rule to check
the value of a field. ie, it is currently not possible to select all parts
that have ModuleCommand and a non-zero CrewCapacity. Since the rule I had
to settle with picks probe cores too, make the workshop disable itself if
it finds itself on a part that cannot hold Kerbals.