Adjust the pad's mass to reflect build progress.
I had completely forgotten that as RocketParts were consumed, the mass of
the vessel would go down, but as the idea is that the consumed parts are
going into the built craft, the overall mass should remain constant. This
"fixes" that. There may be problems with the vessel CoM not updating while
warping, but I don't see a quick solution anywhere, so I'll just wait and
see what happens.

Fixes #43