Change the active vessel forcing.
Force the active vessel to be the spawned vessel (this will take care of
any switch failures on spawn due to KSP being picky) and do so before
starting staging. It seems the NRE in PatchedConicRenderer was caused by
the current vessel not being correct when the orbit driver (?) tries to set
up the vessel.

Fixes #49.