Do an initial pass on balancing part and resource costs.
This includes fixing the masses of the augers: they were about a quarter of
the mass they should be.

Resources are set to something that seems reasonable based on real-world
numbers but using my own markup system.

The price costs are based on their mass with arbitrary (but consistent for
part type) markup: mass * RocketParts cost * markup scale. 1.2 for tanks,
12.8 for detectors, 1.5 for augers, 2.5 for recycling bins and pads, 25 for
the workshop, 20 for smelters.