Extend the tip of the stake's collider 1cm.
If a part's collider doesn't actually collide with the ground, the part
won't consider itself to have ground contact and since the stake is the
only part in the vessel, the vessel won't be landed and KSP will delete it
when the vessel unloads. Blah. Thus extend the tip 1cm past the attach
node.