Implement the core of the survey site tracker.
Survey stakes are added to sites on a celestial body and sites are merged
based on name (taken from the stake) and distance between stakes (<= 200m).
When stakes are removed, the site is split into multiple sites if the stake
was a bridge between distant stake( cluster)s.

The checks aren't particularly optimal, but they happen infrequently so
that shouldn't be a big problem.