Revamp the pad highlight code.
Move the code that actually highlights the pad's part into the builder.
This allows pads to be highlighted as normal, but survey sites to highlight
all the stakes in the site.

The currently selected pad should not be lost on docking events.

Survey stakes use the new highlighting system available as of KSP 0.90. It
makes them a fair bit easier to find.

Highlight state is updated only when the selected pad changes (or highlight
enabling is toggled). This should give a minor fps improvement when the
build window is open.