Handle loading in the Dewarping state.
It seems that sometimes the game can get saved while dewarping and this
causes the pad to get stuck in the dewarping state because the coroutine
does not execute when the game is loaded. Rather than restarting the
coroutine, just assume that it's safe to go directly to the Complete state.