Implement "bounds" stakes.
This gives better control of the position of the built vessel.

As for direction stakes, bounds stakes of the same use (+X, -Y etc) will be
grouped together and their average position used to determine the bound
point.

If only one bound point is given for an axis (eg, either +X or -X, but not
both), the that plane of the vessel's bounding box will touch that point.

If both bound points are given, then the appropriate plane that passes
through the center of the vessel's bounding box will touch that point. This
is in contrast with the origin point being where the vessel's root part
will be placed.

Origin bound stakes have no meaning: they are currently ignored.