Delay the start of workshop activity.
It seems that Unity calls FixedUpdate right after calling Start and thus
workshops could start producing work before pads are ready to accept work
if the pad comes "later" in the vessel tree than the master workshop. This
fixes the NRE I would get in DoWork when switching to a construction vessel
that was currently building something. This may also fix the reports of
background builds not working for some people.