Support building rerooted craft.
With the advent of the root part changer in 0.90, the method to place a
vessel changed slightly. This caused the vessel to have a strange
orientation and location relative to what it should have had based on the
launch transform. Many thanks to Tarheel1999 and beeble42 for finding the
reproduction steps.

I suspect this may be the cause of the odd report I got before 0.90 if
those people were using the root changing mod.