Run through craftVessel.parts instead of ship.Parts.
ship is ShipConstruct and craftVessel is Vessel (created from ship). It may
or may not be a problem now, but it might become one in the future. I
suspect it is a holdover from when I found I had to initialize the vessel
properly to get the correct masses for procedural-ish parts (MFT etc).