Give pilots a use (as surveyors, extending survey range).
My thinking is that pilots are navigators and navigators are not far from
surveyors.

The base range of survey stations has been reduced to 20m (from 100m), but
this is for unmanned stations. Manned stations start at 50m (no skill at
all), 100m for 0-star pilots, up to 2000m for 5-star pilots.