Ensure res_index is always valid.
Going from a part that recycles to more resources to a part that recycles
to fewer resources could result in an invalid resource index. Wrapping the
res_index before getting the resource instead of after ensures it is valid.
Also, though it should not happen, check for 0 resources before entering
the loop. This fixes the problem with the recycler getting stuck in the
transferring state.