Do not null buildCost after vessel release.
This will prevent build cost being calculated every time a pad is loaded
when the pad had built and released a vessel. This was happening because
craftConfig was still valid but buildCost had been nulled. Since
craftConfig is still valid, buildCost should be too, though rebuilding the
same craft causes a full reload so it doesn't matter.