Add a ControlReference part and module.
The part allows for the setting of the vessel control reference ("Control
From Here") in an arbitrary directly without adding an extra command pod or
docking port.

Also, the module adds a "toggle reference" action group that causes the
part to take the vessel reference if it is not already the reference and
remembers the previous reference so it can be restored when toggled again.