Ensure sources and sinks know the worknet.
As part of the spawn process, especially when using the micro-pad, the
worknet for some or all parts may be lost. Since the vessel worknet
rebuilds the network after any vessel modification, simply update each
source and sink with the current worknet while building the network. Fixes
the erroneous 0 productivity in the build window.