Bring back the extending launch clamp module.
It turns out that the stock LaunchClamp module has some bugs that are very
difficult to work around and that cause critical problems for use away from
the launchpad. Thus this is a reimplementation of LaunchClamp that does a
much better job of extending, supporting all the collider types (except
wheel), cloned-segment tower meshes, and spawning over uneven terrain.