Always check both part crew lists and command seats.
KSP's command seat assignment works by setting the command seat's crew
capacity to non-zero (1 for the stock seat), so that is no longer an
indication of whether the part is a command seat or not, so instead always
assume it's both and just combine the two lists. In general, one will be
empty, I imagine.