Add a ForceHeightCheck pseudo-field to launch clamps.
If ForceHeightCheck is true, then the extension of the launch clamp will be
recalculated. Currently, the only way to set it is by hacking the sfs file,
but this is good for "fixing" launch clamps that have somehow been pushed
into the ground by KSP landed vessel logic and are thus twitching due to
PhysX collider logic.