Fix a hang with multi-resource builds.
If one or more required resources were not available, then the work for
the available resources would decrease exponentially as the unavailable
resources were counted for dividing the work but were not actually
taking any portion of the available work. This lead to the kerbalHours
becoming a microscopic number and rounding causing things to never get
to zero.