Handle KIS container masses correctly
KIS containers [correctly] include the mass of any resources in contained
items in the container's mass. However, this causes the build resource
costs to be miscalculated if the resources mass is not subtracted from the
part mass.

The question remaining is whether the contained resource mass should count
toward's the part's dry-mass. For now, it does not.