Remove craft capacities display
While a nice feature, it required the craft resources being found every
time the pad was loaded. This meant the craft config had to be parsed
and the temporary vessel created and destroyed every time, which is
quite wasteful. Instead, a BOM display is planned, and will be even more
informative.