Rework the resource manager
The resource manager now supports exclusion sets and not jumping across
inter-vessel links (eg, KAS resource connector). This is needed for
using the resource manager when building craft as the pad vessel and
attach craft vessel need to be properly separated, and any vessels
linked to the craft vessel not included in the craft vessel. This does
not properly handle vessels docked to the craft vessel (they will be
seen as part of the craft vessel).