Add points to every face they are above
Adding a point to only the first face it is above can cause other faces
the point is above to be treated as final, resulting in the face being
replaced by convex faces later on. This may actually indicate other
logic errors in the face lighting and horizon finding code, but it works
for now.