Implement face splitting
It helps, but there is still a problem. It looks to be when the split
face is in the final set, which implies that perhaps face lighting
should be done in only one pass. Adding edge-touch tests to the
visibility tests stops the edge splitting, but does not fix some faces
being only partially connected.