Delay editing launch clamp part name
This should fix a major incompatibility with Restock. The problem being:

EL creates a clone of the launch clamp (as ELExtendingLaunchClamp)
replacing the LaunchClamp module with the ELExtendingLaunchClamp

ReStock then comes in and replaces the LaunchClamp module on the
original part with the ReStock module.

Then, at build time, EL finds launchClamp1 on the vessel, ass-u-mes its
module is LaunchClamp, edits the part name to be ELExtendingLaunchClamp,
tries to edit the module to point to EL's module but fails because it
looks very specifically for "LaunchClamp"

This results in a broken part.