Clean up craft vessel repositioning
The repositioning is needed while waiting to capture the craft, both in
orbit and landed, to ensure the craft is in a position and orientation
suitable for capture. Survey sites don't actually need repositioning,
but the interface needs the implementation and might as well implement
it properly as I might find a use for it later.