Put non-master branch name in version string
I think one of the reasons I shied away from git-branching is the worry
I'd forget to merge to master before releasing. While this won't prevent
a release tag on a branch, it will at least catch it for many cases (tag
on branch but branch merged into master and currently on master might be
interesting).