Force the landed status of vessels touching pads
It seems that recent versions of KSP have a bug in their landed logic
that causes vessels landed on other (landed) vessels to sometimes not be
landed (particularly if things glitch during scene load, tossing the
upper vessel up and it landing again) but still suborbital. Copying the
landed and splashed flags seems to be sufficient fix it (at least when
several colliders are touching) as the vessel sets its situation
correctly the next frame.