Remove part fx when getting build cost
Something in KSP 1.10 has changed such that particle effects remain
(and active) when the vessel is destroyed the same frame as it was
created. It may be that the effects were always there and they now are
visible. However, with this, they are very much gone (verified using
debug stuff).