Start work on on-demand thumbnail generation
It sort of works, but with a few problems. The main one is the displayed
thumbnails don't update automatically, but it turns out the textures are
not being handled properly so I left that for now as sorting out the
texture handling should sort out the thumbnails too. The other problem
is that because everything is in the current scene, the current lighting
affects the thumbnails.