Get the part selector mostly up and running
Can't actually select a part yet, category icons need a lot of work (and
tooltips), and no info is displayed, but the fundamentals are there.
However, there's a problem with the part icons showing behind other UI
windows that are behind the craft browser. Not sure if the problem is in
Unity, KSP or just my setup (it's entirely possible the problem is there
in the editor, just not noticeable because there are no cousin UI
elements behind the parts list).