Make a start on the part editor
The basic layout is done (simpler than I expected) but with placeholders
for the description and tweakables editors. The description editor needs
multi-line text input (don't know if that will be a separate class or if
I'll tweak UIInputField), and the tweakables editor will be a separate
class when I get to it.