Block attached ground part coroutines
While it turns out to not matter for properly grounded bases, ground
science parts spam the message box when they're not in contact with the
ground, but this could be a problem if the parts are built while still
in flight. It turns out I'd been chasing problems caused by my deployed
parts taking too long to settle to the ground (no local gravity
compensation).