Ensure craft vessel has a valid orbit
This is in the build cost calculation only, but does take care of the
vessel being ripped out from underneath EL while calculating the cost
due to ObT being nan. It looks like this is the same cause for the
breakdown when actually building a survey stake, but that might be a
little trickier to sort out.