Ensure built vessel has a valid orbit
And the orbit driver has a vessel. This fixes the vessel (for a stake,
no idea why that was singled out) being ripped out from under EL and
thus causing all sorts of things to break. The orbit driver needed its
vessel set explicitly because the driver is added before the vessel and
thus has nothing to find.