[survey] Add support for configurable ranges
The ranges for the survey station can be configured on a per-part basis
using the SiteRanges field in the module node. This is just a list of
comma-separated ranges in meters. 0 or negative ranges use the default
range, missing elements (ie, too few in the list) use the default value
for those ranges, excess ranges are ignored, and any parsing error
aborts the list leaving unprocessed ranges at their default value.

Closes #179