[Parts] Add ModuleCargoPart to all active parts
Not bothering with the deprecated parts. The survey stake has
ModuleGroundPart instead of ModuleCargoPart so it can be placed on the
ground (hopefully by any kerbal) even without KIS. The modules are there
regardless of whether KIS is installed or not. This should prevent any
NRE spam attempting to use the parts with cargo modules (though fitting
some of the bigger parts into storage...)

Closes #182