[Resources] Correct energy consumption for ECSmelter recipe
The base data was correct, but there was a lot of confusion, in both the
comments and the math, about just what a kWh is. If *my* math is
correct, even with this correction, smelting ore to metal takes less
power than sintering scrap to rocket parts, so it shouldn't be too bad.